add QMTech XC7K325T board, add seven segment display to daughterboard
