+++ litex_boards/platforms/qmtech_daughterboard.py
@@ -1,4 +1,62 @@
+#
+# This file is part of LiteX-Boards.
+#
+# Copyright (c) 2023 Kazumoto Kojima
+# Copyright (c) 2023 Hans Baier <hansfbaier@gmail.com>
+#
+# SPDX-License-Identifier: BSD-2-Clause
+
 from litex.build.generic_platform import Subsignal, Pins, IOStandard, Misc
+
+# SevenSeg -----------------------------------------------------------------------------------------
+from migen import *
+from migen.genlib.misc import WaitTimer
+
+from litex.soc.interconnect.csr import AutoCSR, CSRStorage
+
+class SevenSeg(Module, AutoCSR):
+    def __init__(self, segs, sels, sys_clk_freq, period=1e-2):
+        self.segs = segs
+        self.sels = sels
+
+        n = len(sels)
+        self._out = CSRStorage(4*n, description="7 Seg LEDs Control.")
+        xdigits = Signal(4*n)
+        select = Signal(n)
+        count = Signal(max=n)
+        table = [
+            0x3f, 0x06, 0x5b, 0x4f,
+            0x66, 0x6d, 0x7d, 0x07,
+            0x7f, 0x6f, 0x77, 0x7c,
+            0x39, 0x5e, 0x79, 0x71
+        ]
+        abcdefg = Signal(8)
+        hexa = Signal(4)
+        cases = {}
+        for i in range(16):
+            cases[i] = abcdefg.eq(table[i])
+
+        self.comb += Case(hexa, cases)
+
+        timer = WaitTimer(int(period*sys_clk_freq/(2*n)))
+        self.submodules += timer
+        self.comb += timer.wait.eq(~timer.done)
+        self.sync += If(timer.done,
+                        If(count == n-1,
+                           count.eq(0),
+                           select.eq(1 << (n-1)),
+                           xdigits.eq(self._out.storage)
+                        ).Else(
+                           count.eq(count + 1),
+                           select.eq(select >> 1),
+                           xdigits.eq(xdigits >> 4)
+                        )
+                     )
+        self.comb += [
+            hexa.eq(xdigits[0:4]),
+            segs.eq(~abcdefg),
+            sels.eq(select)
+        ]
 
 class QMTechDaughterboard:
     """
